chore: moving common classes into src/common
