+++ src/common/atlas/apiClient.ts
@@ -5,7 +5,7 @@ import { ApiClientError } from "./apiClientError.js";
 import { paths, operations } from "./openapi.js";
 import { CommonProperties, TelemetryEvent } from "../../telemetry/types.js";
 import { packageInfo } from "../../helpers/packageInfo.js";
-import logger, { LogId } from "../../logger.js";
+import logger, { LogId } from "../logger.js";
 
 const ATLAS_API_VERSION = "2025-03-12";
