+++ src/telemetry/telemetry.ts
@@ -1,7 +1,7 @@
-import { Session } from "../session.js";
+import { Session } from "../common/session.js";
 import { BaseEvent, CommonProperties } from "./types.js";
-import { UserConfig } from "../config.js";
-import logger, { LogId } from "../logger.js";
+import { UserConfig } from "../common/config.js";
+import logger, { LogId } from "../common/logger.js";
 import { ApiClient } from "../common/atlas/apiClient.js";
 import { MACHINE_METADATA } from "./constants.js";
 import { EventCache } from "./eventCache.js";
