+++ tests/integration/tools/mongodb/mongodbHelpers.ts
@@ -4,7 +4,7 @@ import { fileURLToPath } from "url";
 import fs from "fs/promises";
 import { MongoClient, ObjectId } from "mongodb";
 import { getResponseContent, IntegrationTest, setupIntegrationTest, defaultTestConfig } from "../../helpers.js";
-import { UserConfig } from "../../../../src/config.js";
+import { UserConfig } from "../../../../src/common/config.js";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
